whole bunch of stuff, finished optimizing performance for any size world, made a minimap, etc.

+++ src/SceneMenu.hpp
@@ -4,6 +4,7 @@
 #include "GameEngine.hpp"
 #include "Action.hpp"
 
+#include <SFML/Graphics.hpp>
 #include <string>
 #include <vector>
 #include <chrono>
@@ -13,7 +14,7 @@ class SceneMenu : public Scene
     std::string m_title;
     std::vector<std::string> m_menuStrings;
     sf::Text m_menuText = sf::Text(m_game.assets().getFont("PixelCowboy"));
-    size_t m_selectedMenuIndex = 0;
+    unsigned int m_selectedMenuIndex = 0;
 
     void init();
